requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
colorama==0.4.6
pyfiglet==0.8.post1
python-whois==0.8.0
inquirer==3.1.3
validators==0.21.2
qrcode==7.4.2
pillow==10.0.1 